\name{createCIEweight}
\alias{createCIEweight}

\title{
    Generates a plot of the CIE UV weight function
}

\description{
    nnn
}
\usage{
    createCIEweight()
}
\arguments{
    \item{file}{a connection or a character string giving the name of the
    file to load.}
    \item{feature}{the spectral feature to examine (e.g. CH1, CH2, ...).}
    \item{ylim}{scale the y-axis when PDF and histogram are plotted.}
    \item{method}{choose method, current altrnatives are F, M, N and A.}
    \item{resolution}{specifies the resolution of the histogram.}
}
\details{
    Results are plotted and a message printed.
}
\value{
    Results are plotted and a message printed.
}
\seealso{
    ml package (yet not documented)
}
\author{
    Øystein Godøy (\email{o.godoy@met.no}), METNO/FOU, 25.04.2006 
}
\keyword{}
