\name{plotval}

\alias{plotval}

\title{
    Plot validation results
}

\description{
    Plots collocated validation results.
}

\usage{
    plotval(x, estimate="FC", station=NULL)
}

\arguments{
    \item{x}{Object returned by \link{readuvi}}
    \item{estimate}{Identification of an estimate}
    \item{station}{Short name for station}
}

\value{
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
