\name{readuvi}

\alias{readuvi}

\title{
    Read UV index estimates
}

\description{
    Reads ASCII files containing UV index estimates generated using
    libRadtran at the Norwegian Meteorological institute. 
}

\usage{
    readuvi(filename)
}

\arguments{
    \item{filename}{Collocation file to read}
}

\value{
    \link{list} with validtime and data as \link{data.frame}
}

\seealso{
    \link{readuvival}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
