\name{readuvi}

\alias{readuvi}

\title{
    Read UV index estimates
}

\description{
    Reasds ASCII files containing collocated UV index estimates and
    observations. Observations are collected from NRPAs observation
    network and stations are identified using short names.
}

\usage{
}

\arguments{
    \item{filename}{Collocation file to read}
}

\value{
    \link{data.frame}
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
