\name{validateuv}

\alias{validateuv}

\title{
    Create a validation plot
}

\description{
    Creates a validation plot of estimated UV index against observed.
}

\usage{
}

\arguments{
    \item{filename}{file to read}
}

\value{
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
