\name{readuvival}

\alias{readuvival}

\title{
    Read validation files for UV index estimates
}

\description{
    Reads ASCII files containing collocated UV index estimates and
    observations. Observations are collected from NRPAs observation
    network and stations are identified using short names.
}

\usage{
}

\arguments{
    \item{filename}{Collocation file to read}
}

\value{
    \link{data.frame}
}

\seealso{
    \link{readuvi}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
